+++ src/MainWindow.xaml.cs
@@ -47,7 +47,7 @@ namespace ZeroMix
 
     public partial class MainWindow : Wpf.Ui.Controls.FluentWindow, ZeroMix.Plugins.IZeroMixHost
     {
-        private const string CURRENT_VERSION = "7.4.1";
+        private const string CURRENT_VERSION = "7.5.0";
 
         // Windows API for Taskbar transparency
         [DllImport("user32.dll", SetLastError = true)]